Fix filter panel rendering and split into component files
Move filter panel positioning from a CSS class to inline styles (matching
the AdvanceTimePopover pattern) so position:fixed applies correctly via the
React portal. Remove the palette guard so the panel renders whenever the
campaign is loaded. Split FilterChip, TagEditor, and DateEditor out of
filter-panel.tsx into their own files per reviewer request.
