+++ desktop/src/renderer/timeline/filter/filter-panel.css
@@ -1,16 +1,3 @@
-.filter-panel {
-  position: fixed;
-  bottom: 50px;
-  left: 0;
-  right: 0;
-  background: var(--theme-panel, #2d3d2a);
-  border-top: 1px solid var(--theme-border, #3a3a30);
-  padding: 8px 12px;
-  z-index: 200;
-  max-height: 200px;
-  overflow-y: auto;
-}
-
 .filter-bar {
   display: flex;
   flex-wrap: wrap;
